refactor: move display management to separate class
