+++ p2p-spike/robot/RobotClient.ts
@@ -1,5 +1,6 @@
-import type { Metrics, Point, Robot } from "./types";
+import type { Metrics, Robot, ServerMetrics } from "./types";
 import PeerConnection from "./PeerConnection";
+import DisplayManager from "./DisplayManager";
 
 class RobotClient {
   peerConnection: PeerConnection;
@@ -7,15 +8,15 @@ class RobotClient {
   robotId: string | null = null;
 
   robots: Map<string, Robot> = new Map();
-  worldSize = { width: 1000, height: 1000 };
-  robotSize = { width: 50, height: 50 };
   robotSpeed = 200;
 
   inputs = { w: false, s: false, a: false, d: false };
   // inputSequence = 0;
   // inputBuffer: Map<number, any> = new Map();
 
-  clientMetrics = {
+  displayManager: DisplayManager;
+
+  clientMetrics: Metrics = {
     startTime: Date.now(),
     totalFrames: 0,
     frameTimes: [] as number[],
@@ -26,85 +27,22 @@ class RobotClient {
     connectionTime: 0,
     averageFPS: 0,
     networkStats: { packetsLost: 0, roundTripTimes: [] as number[], jitter: 0 },
-    serverMetrics: {} as Metrics,
+    serverMetrics: {} as ServerMetrics,
   };
 
-  canvas: HTMLCanvasElement;
-  ctx: CanvasRenderingContext2D;
-  lastRenderTime = 0;
-
-  // HTML Elements
-  statusEl: HTMLElement;
-  playerCountEl: HTMLElement;
-  latencyEl: HTMLElement;
-  serverTickEl: HTMLElement;
-  correctionCountEl: HTMLElement;
-  avgDivergenceEl: HTMLElement;
-  inputLagEl: HTMLElement;
-  inputSequenceEl: HTMLElement;
-  lastCorrectionEl: HTMLElement;
-  fpsEl: HTMLElement;
-  serverTickRateEl: HTMLElement;
-  networkThroughputEl: HTMLElement;
-  serverMemoryEl: HTMLElement;
-  totalMessagesEl: HTMLElement;
-  uptimeEl: HTMLElement;
-  jitterEl: HTMLElement;
-  infoPanel: HTMLElement;
-  serverPanel: HTMLElement;
-  controlsPanel: HTMLElement;
-  metricsVisible = true;
-
   constructor() {
-    this.clientId = this.generateClientId();
-    this.peerConnection = new PeerConnection(
-      this.clientId,
-      this.handlePeerMessage,
-    );
+    this.peerConnection = new PeerConnection(this.handlePeerMessage);
+    this.clientId = this.peerConnection.clientId;
+    this.displayManager = new DisplayManager();
 
-    this.canvas = document.getElementById("gameCanvas") as HTMLCanvasElement;
-    this.ctx = this.canvas.getContext("2d")!;
-    this.statusEl = document.getElementById("status")!;
-    this.playerCountEl = document.getElementById("playerCount")!;
-    this.latencyEl = document.getElementById("latency")!;
-    this.serverTickEl = document.getElementById("serverTick")!;
-    this.correctionCountEl = document.getElementById("correctionCount")!;
-    this.avgDivergenceEl = document.getElementById("avgDivergence")!;
-    this.inputLagEl = document.getElementById("inputLag")!;
-    this.inputSequenceEl = document.getElementById("inputSequence")!;
-    this.lastCorrectionEl = document.getElementById("lastCorrection")!;
-    this.fpsEl = document.getElementById("fps")!;
-    this.serverTickRateEl = document.getElementById("serverTickRate")!;
-    this.networkThroughputEl = document.getElementById("networkThroughput")!;
-    this.serverMemoryEl = document.getElementById("serverMemory")!;
-    this.totalMessagesEl = document.getElementById("totalMessages")!;
-    this.uptimeEl = document.getElementById("uptime")!;
-    this.jitterEl = document.getElementById("jitter")!;
-    this.infoPanel = document.getElementById("info")!;
-    this.serverPanel = document.getElementById("serverStats")!;
-    this.controlsPanel = document.getElementById("controls")!;
-
-    this.setupCanvas();
+    this.displayManager.setupCanvas();
     this.setupInputHandlers();
-    this.startRenderLoop();
+    this.displayManager.startRenderLoop(
+      this.robots,
+      this.robotId ?? "",
+      this.clientMetrics,
+    );
     this.startMetricsCollection();
-
-    // Create or connect to a peer
-    this.peerConnection.peer.on("open", (id) => {
-      console.log(`My peer ID is: ${id}`);
-      // Connect to another peer (for example using a fixed peer ID for simplicity)
-      this.peerConnection.connectToPeer("some-peer-id"); // Replace 'some-peer-id' with the actual peer ID
-    });
-
-    this.peerConnection.peer.on("connection", (conn) => {
-      console.log("Connected to another peer");
-      this.peerConnection.connection = conn;
-      this.peerConnection.setupConnectionHandlers();
-    });
-  }
-
-  generateClientId(): string {
-    return `client-${Math.random().toString(36).substring(2, 9)}`;
   }
 
   handlePeerMessage(data: any) {
@@ -139,17 +77,17 @@ class RobotClient {
   handleInit(data: any) {
     this.clientId = data.clientId;
     this.robotId = data.robotId;
-    this.worldSize = data.worldSize;
+    this.displayManager.setWorldsize(data.worldSize);
 
     data.robots.forEach((robot: Robot) => {
       this.robots.set(robot.id, { ...robot });
     });
-    this.playerCountEl.textContent = data.robots.length;
+    this.displayManager.updatePlayerCount(data.robots.length);
     console.log(`Initialized with ${data.robots.length} robots`);
   }
 
   handleGameState(data: any) {
-    this.serverTickEl.textContent = data.sequence;
+    this.displayManager.updateServerTick(data.sequence);
 
     data.robots.forEach((robotData: any) => {
       const robot = this.robots.get(robotData.id);
@@ -160,18 +98,18 @@ class RobotClient {
       }
     });
 
-    this.playerCountEl.textContent = data.robots.length;
+    this.displayManager.updatePlayerCount(data.robots.length);
   }
 
   handleRobotJoined(data: any) {
     this.robots.set(data.id, { ...data });
-    this.playerCountEl.textContent = this.robots.size.toString();
+    this.displayManager.updatePlayerCount(this.robots.size);
     console.log(`Robot ${data.id} joined`);
   }
 
   handleRobotLeft(data: any) {
     this.robots.delete(data.robotId);
-    this.playerCountEl.textContent = this.robots.size.toString();
+    this.displayManager.updatePlayerCount(this.robots.size);
     console.log(`Robot ${data.robotId} left`);
   }
 
@@ -200,7 +138,6 @@ class RobotClient {
 
   handleServerMetrics(data: any) {
     this.clientMetrics.serverMetrics = data;
-    this.updateServerMetricsDisplay();
   }
 
   startMetricsCollection() {
@@ -225,45 +162,14 @@ class RobotClient {
   }
 
   updateMetricsDisplay() {
-    if (this.fpsEl)
-      this.fpsEl.textContent = this.clientMetrics.averageFPS.toString() || "0";
-    if (this.totalMessagesEl)
-      this.totalMessagesEl.textContent =
-        this.clientMetrics.messagesReceived.toString();
-    if (this.jitterEl) {
-      this.jitterEl.textContent =
-        Math.round(this.clientMetrics.networkStats.jitter * 10) / 10 + "ms";
-    }
-    if (this.uptimeEl) {
-      const uptime = Date.now() - this.clientMetrics.startTime;
-      const seconds = Math.floor(uptime / 1000) % 60;
-      const minutes = Math.floor(uptime / 60000);
-      this.uptimeEl.textContent = `${minutes}:${seconds.toString().padStart(2, "0")}`;
-    }
-  }
-
-  updateServerMetricsDisplay() {
-    const serverMetrics = this.clientMetrics.serverMetrics;
-    if (!serverMetrics) return;
-
-    if (this.serverTickRateEl) {
-      this.serverTickRateEl.textContent = `${serverMetrics.actualTickRate}/${serverMetrics.tickRate}`;
-    }
-    if (this.networkThroughputEl) {
-      this.networkThroughputEl.textContent = `${serverMetrics.messagesPerSecond}/s`;
-    }
-    if (this.serverMemoryEl) {
-      this.serverMemoryEl.textContent = `${serverMetrics.memoryUsageMB}MB`;
-    }
-  }
-
-  showCorrectionIndicator(type: string) {
-    this.lastCorrectionEl.textContent = type.toUpperCase();
-    this.lastCorrectionEl.classList.add("correction-flash");
-
-    setTimeout(() => {
-      this.lastCorrectionEl.classList.remove("correction-flash");
-    }, 500);
+    this.displayManager.updateFPS(this.clientMetrics.averageFPS ?? 0);
+    this.displayManager.updateTotalMessages(
+      this.clientMetrics.messagesReceived,
+    );
+    this.displayManager.updateJitter(
+      Math.round(this.clientMetrics.networkStats.jitter * 10) / 10,
+    );
+    this.displayManager.updateUptime(this.clientMetrics.startTime);
   }
 
   sendClientState() {
@@ -281,15 +187,7 @@ class RobotClient {
     });
   }
 
-  setupCanvas() {
-    this.canvas.width = this.worldSize.width;
-    this.canvas.height = this.worldSize.height;
-
-    window.addEventListener("resize", () => {
-      this.canvas.width = this.worldSize.width;
-      this.canvas.height = this.worldSize.height;
-    });
-  }
+  setupCanvas() {}
 
   setupInputHandlers() {
     document.addEventListener("keydown", (e) => {
@@ -314,7 +212,7 @@ class RobotClient {
           this.resetStats();
           return;
         case "m":
-          this.toggleMetrics();
+          this.displayManager.toggleMetrics();
           return;
       }
     }
@@ -354,14 +252,16 @@ class RobotClient {
     clientRobotState.position.x = Math.max(
       0,
       Math.min(
-        this.worldSize.width - this.robotSize.width,
+        this.displayManager.worldSize.width -
+          this.displayManager.robotSize.width,
         clientRobotState.position.x,
       ),
     );
     clientRobotState.position.y = Math.max(
       0,
       Math.min(
-        this.worldSize.height - this.robotSize.height,
+        this.displayManager.worldSize.height -
+          this.displayManager.robotSize.height,
         clientRobotState.position.y,
       ),
     );
@@ -380,8 +280,6 @@ class RobotClient {
   }
 
   resetStats() {
-    this.lastCorrectionEl.textContent = "None";
-
     // Reset client metrics
     this.clientMetrics.totalFrames = 0;
     this.clientMetrics.frameTimes = [];
@@ -393,130 +291,11 @@ class RobotClient {
     console.log("Stats reset");
   }
 
-  toggleMetrics() {
-    this.metricsVisible = !this.metricsVisible;
-    const display = this.metricsVisible ? "block" : "none";
-
-    if (this.infoPanel) this.infoPanel.style.display = display;
-    if (this.serverPanel) this.serverPanel.style.display = display;
-    if (this.controlsPanel) this.controlsPanel.style.display = display;
-
-    console.log(`Metrics panels ${this.metricsVisible ? "shown" : "hidden"}`);
-  }
-
   sendMessage(message: any) {
     if (this.peerConnection && this.peerConnection.connected) {
       const messageStr = JSON.stringify(message);
       this.clientMetrics.bytesSent += messageStr.length;
       this.peerConnection.send(messageStr);
-    }
-  }
-
-  startRenderLoop() {
-    const render = (timestamp: number) => {
-      this.render(timestamp);
-      requestAnimationFrame(render);
-    };
-    requestAnimationFrame(render);
-  }
-  render(timestamp: number) {
-    // Track frame timing
-    if (this.lastRenderTime > 0) {
-      const frameTime = timestamp - this.lastRenderTime;
-      this.clientMetrics.frameTimes.push(frameTime);
-
-      if (this.clientMetrics.frameTimes.length > 60) {
-        this.clientMetrics.frameTimes.shift();
-      }
-    }
-    this.clientMetrics.totalFrames++;
-
-    this.ctx.fillStyle = "#000";
-    this.ctx.fillRect(0, 0, this.canvas.width, this.canvas.height);
-
-    this.drawGrid();
-
-    for (const [robotId, robot] of this.robots) {
-      if (robotId === this.robotId) {
-        this.drawOwnRobot(robot);
-      } else {
-        this.drawOtherRobot(robot);
-      }
-    }
-
-    this.lastRenderTime = timestamp;
-  }
-
-  drawGrid() {
-    this.ctx.strokeStyle = "#333";
-    this.ctx.lineWidth = 1;
-
-    for (let x = 0; x <= this.worldSize.width; x += 100) {
-      this.ctx.beginPath();
-      this.ctx.moveTo(x, 0);
-      this.ctx.lineTo(x, this.worldSize.height);
-      this.ctx.stroke();
-    }
-
-    for (let y = 0; y <= this.worldSize.height; y += 100) {
-      this.ctx.beginPath();
-      this.ctx.moveTo(0, y);
-      this.ctx.lineTo(this.worldSize.width, y);
-      this.ctx.stroke();
-    }
-  }
-
-  drawOwnRobot(robot: Robot) {
-    this.drawRobotAt(robot.position, robot.rotation, "#ff4444", "YOU", false);
-  }
-
-  drawOtherRobot(robot: Robot) {
-    this.drawRobotAt(robot.position, robot.rotation, "#44ff44", "OTHER", false);
-  }
-
-  drawRobotAt(
-    position: Point,
-    rotation: number,
-    color: string,
-    label: string,
-    secondary: boolean,
-  ) {
-    this.ctx.save();
-    this.ctx.translate(
-      position.x + this.robotSize.width / 2,
-      position.y + this.robotSize.height / 2,
-    );
-    this.ctx.rotate((rotation * Math.PI) / 180);
-
-    this.ctx.fillStyle = color;
-    this.ctx.fillRect(
-      -this.robotSize.width / 2,
-      -this.robotSize.height / 2,
-      this.robotSize.width,
-      this.robotSize.height,
-    );
-
-    this.ctx.fillStyle = "#fff";
-    this.ctx.fillRect(this.robotSize.width / 2 - 5, -2, 10, 4);
-
-    this.ctx.restore();
-
-    this.ctx.font = "10px monospace";
-    this.ctx.textAlign = "center";
-    if (secondary) {
-      this.ctx.fillStyle = "#aaa";
-      this.ctx.fillText(
-        label,
-        position.x + this.robotSize.width / 2,
-        position.y + 62,
-      );
-    } else {
-      this.ctx.fillStyle = "#fff";
-      this.ctx.fillText(
-        label,
-        position.x + this.robotSize.width / 2,
-        position.y - 5,
-      );
     }
   }
 }
