+++ p2p-spike/robot/server.ts
@@ -3,7 +3,6 @@ import { PeerServer } from "peer";
 export const PORT = 9000;
 
 const peerServer = PeerServer({ port: PORT, path: "/" });
-peerServer.listen();
 
 console.log(`WebRTC Connection Server Running on Port ${PORT}`);
